refactor: extract notify() into EventNotifierService (#385)
EventService.notify() now delegates to the new concrete service. No
behavior change; first step toward removing the EventService/StateService
circular dependency.
